Small fixes (indents, comparisons) and moved all cog-related calculations into one file (#23)
* fixed indentation in plotting.py

* fixes to analysis.py: removed cog calculation, cleaned up comparisons

* moved cog_cell to centreofgravity.py

* small fixes to tracking.py (comparisons)

* removed cog_cell from Routine Listings in analysis.py

* small fix to feature_detection.py (added empty regions_old dictionary)

* fix to __init__.py for moving cog_cell function to centreofgravity module
