+++ tobac/analysis.py
@@ -24,8 +24,6 @@ cell_statistics(input_cubes, track, mask, aggregators, cell,
                 output_path, output_name, width, z_coordinate,
                 dimensions, **kwargs)
 
-cog_cell(cell, Tracks, M_total, M_liquid, M_frozen, Mask, savedir)
-
 lifetime_histogram(Track, bin_edges, density, return_values)
 
 haversine(lat1, lon1, lat2, lon2)
@@ -242,7 +240,7 @@ def cell_statistics(input_cubes,track,mask,aggregators,cell,output_path='./',out
                 #remove all collapsed coordinates (x and y dim, scalar now) and keep only time as all these coordinates are useless
                 for coordinate in cube_collapsed.coords():
                     if not cube_collapsed.coord_dims(coordinate):
-                        if coordinate.name() is not 'time':
+                        if coordinate.name() != 'time':
                             cube_collapsed.remove_coord(coordinate)
                 logging.debug(str(cube_collapsed))
                 cubes_profile[aggregator.name()].append(cube_collapsed)
@@ -264,74 +262,6 @@ def cell_statistics(input_cubes,track,mask,aggregators,cell,output_path='./',out
         os.makedirs(os.path.join(output_path,output_name,aggregator.name()),exist_ok=True)
         savefile=os.path.join(output_path,output_name,aggregator.name(),output_name+'_'+ aggregator.name()+'_'+str(int(cell))+'.nc')
         save(cubes_profile[aggregator.name()],savefile)
-
-
-def cog_cell(cell,Tracks=None,M_total=None,M_liquid=None,
-             M_frozen=None,
-             Mask=None,
-             savedir=None):
-    '''
-    Parameters
-    ----------
-    cell : int
-        Integer id of cell to create masked cube for output.
-
-    Tracks : optional
-        Default is None.
-
-    M_total : subset of cube, optional
-        Default is None.
-
-    M_liquid : subset of cube, optional
-        Default is None.
-
-    M_frozen : subset of cube, optional
-        Default is None.
-
-    savedir : str
-        Default is None.
-
-    Returns
-    -------
-    None
-
-    Notes
-    -----
-    unsure about anything
-    needs a short summary
-    '''
-    
-    from iris import Constraint
-    logging.debug('Start calculating COG for '+str(cell))
-    Track=Tracks[Tracks['cell']==cell]
-    constraint_time=Constraint(time=lambda cell: Track.head(1)['time'].values[0] <= cell <= Track.tail(1)['time'].values[0])
-    M_total_i=M_total.extract(constraint_time)
-    M_liquid_i=M_liquid.extract(constraint_time)
-    M_frozen_i=M_frozen.extract(constraint_time)
-    Mask_i=Mask.extract(constraint_time)
-
-    savedir_cell=os.path.join(savedir,'cells',str(int(cell)))
-    os.makedirs(savedir_cell,exist_ok=True)
-    savefile_COG_total_i=os.path.join(savedir_cell,'COG_total'+'_'+str(int(cell))+'.h5')
-    savefile_COG_liquid_i=os.path.join(savedir_cell,'COG_liquid'+'_'+str(int(cell))+'.h5')
-    savefile_COG_frozen_i=os.path.join(savedir_cell,'COG_frozen'+'_'+str(int(cell))+'.h5')
-    
-    Tracks_COG_total_i=calculate_cog(Track,M_total_i,Mask_i)
-#   Tracks_COG_total_list.append(Tracks_COG_total_i)
-    logging.debug('COG total loaded for ' +str(cell))
-    
-    Tracks_COG_liquid_i=calculate_cog(Track,M_liquid_i,Mask_i)
-#   Tracks_COG_liquid_list.append(Tracks_COG_liquid_i)
-    logging.debug('COG liquid loaded for ' +str(cell))
-    Tracks_COG_frozen_i=calculate_cog(Track,M_frozen_i,Mask_i)
-#   Tracks_COG_frozen_list.append(Tracks_COG_frozen_i)
-    logging.debug('COG frozen loaded for ' +str(cell))
-    
-    Tracks_COG_total_i.to_hdf(savefile_COG_total_i,'table')
-    Tracks_COG_liquid_i.to_hdf(savefile_COG_liquid_i,'table')
-    Tracks_COG_frozen_i.to_hdf(savefile_COG_frozen_i,'table')
-    logging.debug('individual COG calculated and saved to '+ savedir_cell)
-
 
 def lifetime_histogram(Track,bin_edges=np.arange(0,200,20),density=False,return_values=False):
     '''
