+++ tobac/tracking.py
@@ -172,7 +172,7 @@ def linking_trackpy(features,field_in,dt,dxy,
     features_linking=deepcopy(features)
     
     
-    if method_linking is 'random':
+    if method_linking == 'random':
 #     link features into trajectories:
         trajectories_unfiltered = tp.link(features_linking, 
                                search_range=search_range, 
@@ -182,7 +182,7 @@ def linking_trackpy(features,field_in,dt,dxy,
                                adaptive_step=adaptive_step,adaptive_stop=adaptive_stop,
                                neighbor_strategy='KDTree', link_strategy='auto'
                                )
-    elif method_linking is 'predict':
+    elif method_linking == 'predict':
 
         pred = tp.predict.NearestVelocityPredict(span=1)
         trajectories_unfiltered = pred.link_df(features_linking, search_range=search_range, memory=memory,
